make assert partial
