+++ src/TestFramework/MSTest.Core/Assertions/Assert.Boolean.cs
@@ -1,0 +1,272 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+#if !NETCOREAPP3_0_OR_GREATER && !NET6_0_OR_GREATER
+#define HIDE_MESSAGELESS_IMPLEMENTATION
+#endif
+
+namespace Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Threading.Tasks;
+
+/// <summary>
+/// A collection of helper classes to test various conditions within
+/// unit tests. If the condition being tested is not met, an exception
+/// is thrown.
+/// </summary>
+public sealed partial class Assert
+{
+
+#if HIDE_MESSAGELESS_IMPLEMENTATION
+    /// <summary>
+    /// Tests whether the specified condition is true and throws an exception
+    /// if the condition is false.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be true.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is false.
+    /// </exception>
+    public static void IsTrue([DoesNotReturnIf(false)] bool condition)
+    {
+        IsTrue(condition, string.Empty, null);
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is true and throws an exception
+    /// if the condition is false.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be true.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is false.
+    /// </exception>
+    public static void IsTrue([DoesNotReturnIf(false)] bool? condition)
+    {
+        IsTrue(condition, string.Empty, null);
+    }
+#endif
+
+    /// <summary>
+    /// Tests whether the specified condition is true and throws an exception
+    /// if the condition is false.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be true.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is false. The message is shown in test results.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is false.
+    /// </exception>
+    public static void IsTrue([DoesNotReturnIf(false)] bool condition,
+        [CallerArgumentExpression("condition")] string message = null)
+    {
+        IsTrue(condition, message, null);
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is true and throws an exception
+    /// if the condition is false.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be true.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is false. The message is shown in test results.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is false.
+    /// </exception>
+    public static void IsTrue([DoesNotReturnIf(false)] bool? condition,
+        [CallerArgumentExpression("condition")] string message = null)
+    {
+        IsTrue(condition, message, null);
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is true and throws an exception
+    /// if the condition is false.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be true.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is false. The message is shown in test results.
+    /// </param>
+    /// <param name="parameters">
+    /// An array of parameters to use when formatting <paramref name="message"/>.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is false.
+    /// </exception>
+    public static void IsTrue([DoesNotReturnIf(false)] bool condition,
+        [CallerArgumentExpression("condition")] string message = null, params object[] parameters)
+    {
+        if (!condition)
+        {
+            ThrowAssertFailed("Assert.IsTrue", BuildUserMessage(message, parameters));
+        }
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is true and throws an exception
+    /// if the condition is false.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be true.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is false. The message is shown in test results.
+    /// </param>
+    /// <param name="parameters">
+    /// An array of parameters to use when formatting <paramref name="message"/>.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is false.
+    /// </exception>
+    public static void IsTrue([DoesNotReturnIf(false)] bool? condition,
+        [CallerArgumentExpression("condition")] string message = null, params object[] parameters)
+    {
+        if (condition is false or null)
+        {
+            ThrowAssertFailed("Assert.IsTrue", BuildUserMessage(message, parameters));
+        }
+    }
+
+#if HIDE_MESSAGELESS_IMPLEMENTATION
+    /// <summary>
+    /// Tests whether the specified condition is false and throws an exception
+    /// if the condition is true.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be false.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is true.
+    /// </exception>
+    public static void IsFalse([DoesNotReturnIf(true)] bool condition)
+    {
+        IsFalse(condition, string.Empty, null);
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is false and throws an exception
+    /// if the condition is true.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be false.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is true.
+    /// </exception>
+    public static void IsFalse([DoesNotReturnIf(true)] bool? condition)
+    {
+        IsFalse(condition, string.Empty, null);
+    }
+#endif
+
+    /// <summary>
+    /// Tests whether the specified condition is false and throws an exception
+    /// if the condition is true.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be false.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is true. The message is shown in test results.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is true.
+    /// </exception>
+    public static void IsFalse([DoesNotReturnIf(true)] bool condition,
+        [CallerArgumentExpression("condition")] string message = null)
+    {
+        IsFalse(condition, message, null);
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is false and throws an exception
+    /// if the condition is true.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be false.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is true. The message is shown in test results.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is true.
+    /// </exception>
+    public static void IsFalse([DoesNotReturnIf(true)] bool? condition,
+        [CallerArgumentExpression("condition")] string message = null)
+    {
+        IsFalse(condition, message, null);
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is false and throws an exception
+    /// if the condition is true.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be false.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is true. The message is shown in test results.
+    /// </param>
+    /// <param name="parameters">
+    /// An array of parameters to use when formatting <paramref name="message"/>.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is true.
+    /// </exception>
+    public static void IsFalse([DoesNotReturnIf(true)] bool condition,
+        [CallerArgumentExpression("condition")] string message = null, params object[] parameters)
+    {
+        if (condition)
+        {
+            ThrowAssertFailed("Assert.IsFalse", BuildUserMessage(message, parameters));
+        }
+    }
+
+    /// <summary>
+    /// Tests whether the specified condition is false and throws an exception
+    /// if the condition is true.
+    /// </summary>
+    /// <param name="condition">
+    /// The condition the test expects to be false.
+    /// </param>
+    /// <param name="message">
+    /// The message to include in the exception when <paramref name="condition"/>
+    /// is true. The message is shown in test results.
+    /// </param>
+    /// <param name="parameters">
+    /// An array of parameters to use when formatting <paramref name="message"/>.
+    /// </param>
+    /// <exception cref="AssertFailedException">
+    /// Thrown if <paramref name="condition"/> is true.
+    /// </exception>
+    public static void IsFalse([DoesNotReturnIf(true)] bool? condition,
+        [CallerArgumentExpression("condition")] string message = null, params object[] parameters)
+    {
+        if (condition is true or null)
+        {
+            ThrowAssertFailed("Assert.IsFalse", BuildUserMessage(message, parameters));
+        }
+    }
+}
